Replaced (mostly) pyre SQL with GORM
Verdict: gorm is unpredictable sometimes, it adds a dependency to the project, so it can get out of control more easily. At the same time the way I did pyre SQL is wrong as i didn't sanatyze correctly.
Lesson to learn: never use fmt.Sprintf! Not safe as it allows SQL injections. Should use db driver way with ? or $ to do so. May be try sqlx lib...